// src/generated/prisma/client/enums.ts
/* !!! This is code generated by Prisma. Do not edit directly. !!! */
/* eslint-disable */
// biome-ignore-all lint: generated file
// @ts-nocheck 
/*
* This file exports all enum related types from the schema.
*
* 🟢 You can import this file directly.
*/

export const Role = {
  USER: 'USER',
  ADMIN: 'ADMIN',
  VOLUNTEER: 'VOLUNTEER'
} as const

export type Role = (typeof Role)[keyof typeof Role]


export const Status = {
  BLOCKED: 'BLOCKED',
  DELETED: 'DELETED',
  ACTIVE: 'ACTIVE'
} as const

export type Status = (typeof Status)[keyof typeof Status]


export const BloodGroup = {
  A_POSITIVE: 'A_POSITIVE',
  A_NEGATIVE: 'A_NEGATIVE',
  B_POSITIVE: 'B_POSITIVE',
  B_NEGATIVE: 'B_NEGATIVE',
  AB_POSITIVE: 'AB_POSITIVE',
  AB_NEGATIVE: 'AB_NEGATIVE',
  O_POSITIVE: 'O_POSITIVE',
  O_NEGATIVE: 'O_NEGATIVE'
} as const

export type BloodGroup = (typeof BloodGroup)[keyof typeof BloodGroup]


export const EmergencyType = {
  MEDICAL: 'MEDICAL',
  FIRE: 'FIRE',
  ACCIDENT: 'ACCIDENT',
  FLOOD: 'FLOOD',
  CRIME: 'CRIME',
  OTHER: 'OTHER'
} as const

export type EmergencyType = (typeof EmergencyType)[keyof typeof EmergencyType]


export const EmergencyStatus = {
  PENDING: 'PENDING',
  ACCEPTED: 'ACCEPTED',
  IN_PROGRESS: 'IN_PROGRESS',
  RESOLVED: 'RESOLVED',
  CANCELLED: 'CANCELLED'
} as const

export type EmergencyStatus = (typeof EmergencyStatus)[keyof typeof EmergencyStatus]


export const VolunteerSkill = {
  FIRST_AID: 'FIRST_AID',
  FIREFIGHTING: 'FIREFIGHTING',
  RESCUE: 'RESCUE',
  MEDICAL: 'MEDICAL',
  FLOOD_RESPONSE: 'FLOOD_RESPONSE',
  GENERAL: 'GENERAL'
} as const

export type VolunteerSkill = (typeof VolunteerSkill)[keyof typeof VolunteerSkill]


export const PaymentMethod = {
  STRIPE: 'STRIPE'
} as const

export type PaymentMethod = (typeof PaymentMethod)[keyof typeof PaymentMethod]


export const PaymentStatus = {
  PENDING: 'PENDING',
  PAID: 'PAID',
  FAILED: 'FAILED',
  REFUNDED: 'REFUNDED'
} as const

export type PaymentStatus = (typeof PaymentStatus)[keyof typeof PaymentStatus]